requirements.txt
asgiref==3.11.0
dj-database-url==3.1.0
Django==5.2.8
django-cors-headers==4.9.0
djangorestframework==3.16.1
djangorestframework_simplejwt==5.5.1
Faker==40.1.2
mysqlclient==2.2.7
pillow==12.0.0
PyJWT==2.10.1
python-dotenv==1.2.1
sqlparse==0.5.4
tzdata==2025.2
